* fixed lots of gcc warnings (mostly signed <-> unsigned conversions)
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@5015 37977431-3df6-0310-b722-df95706aa16b

+++ rts/System/FileSystem/SimpleParser.cpp
@@ -81,7 +81,7 @@ vector<string> CSimpleParser::Tokenize(const string& line, int minWords)
 			break;
 		}
 		string word;
-		if ((minWords > 0) && (words.size() >= minWords)) {
+		if ((minWords > 0) && ((int)words.size() >= minWords)) {
 			word = line.substr(start);
 			// strip trailing whitespace
 			string::size_type pos = word.find_last_not_of(" \t");
